cambios a la autenticacion. faltaria modificar el logout y cambiar contraseña, ademas de arreglar el delete user

+++ harmony-backend/controllers/userController.ts
@@ -1,22 +1,22 @@
 import {FastifyInstance, FastifyRequest} from 'fastify'
 import server, {logger} from "../server";
-import {UserPersistence} from '../persistence/userPersistence';
-import {AuthService} from '../service/authService';
+import {AuthService, UserAuth} from '../service/authService';
 import {UserService} from '../service/userService';
 import {parseCreateUserRequest} from '../models/createUserRequest';
 import {parseChangeIconRequest} from '../models/changeIconRequest';
-import {handleError, parseId} from '../utils';
+import {handleError, parseId, parseJWT} from '../utils';
+import {UserResponse} from "@supabase/supabase-js";
 
 const BASE_URL = '/api/user'
 
 export default async function userController(fastify: FastifyInstance, opts: any) {
-    
-    server.get(BASE_URL + '/:id', async (req: FastifyRequest<{Params: {id: number}}> , rep) => {
+
+    server.get(BASE_URL + '/:id', async (req: FastifyRequest<{ Params: { id: number } }>, rep) => {
         const id = req.params.id;
         try {
             parseId(id)
             logger.info("Getting user with id: " + id);
-            return await UserPersistence.getUserWithID(id);
+            return await UserService.getUserById(id);
         } catch (error: any) {
             logger.error(error)
             return handleError(error, rep)
@@ -27,43 +27,54 @@ export default async function userController(fastify: FastifyInstance, opts: any
         try {
             const request = parseCreateUserRequest(req.body)
             logger.info("Signing up user with email: " + request.email);
-            const data = await AuthService.signUpNewUser(request);
-            if(data.user?.id == undefined){ 
-                //Creo que aca no entra nunca, porque catchea el error que tiro desde la función
+            const userData = await UserService.createUser(request.email, request.name);
+            const data = await AuthService.signUpNewUser(request, userData.id);
+            if (data.access_token == null) {
                 return rep
                     .code(409)
                     .send()
             }
-            return await UserService.createUser(request.email, request.name, data.user.id);
+            return data
         } catch (error: any) {
             logger.error(error)
             return handleError(error, rep)
         }
-        
+
     });
 
-    server.delete(BASE_URL + '/:id', async (req: FastifyRequest<{Params: {id: number}}> , rep) => {
+
+    // TODO FIX DELETE USER
+    server.delete(BASE_URL + '/:id', async (req: FastifyRequest<{ Params: { id: number } }>, rep) => {
         const id = req.params.id;
-        try{
+        try {
             parseId(id)
-            const auth_id = await UserService.getAuthId(id);
-            if (auth_id == null){
-                //Creo que no entra aca porque catchea el error que tiro en el userService
-                logger.info("User with ID " + id + " does not exists, so it can't be deleted")
+            const userAuth: UserAuth = parseJWT(req.headers.authorization || "")
+            if (userAuth.access_token == null) {
+                return rep
+                    .code(401)
+                    .send()
+            }
+            const supabaseReponse: UserResponse = await AuthService.getLoggedUser(userAuth.access_token)
+            logger.info(supabaseReponse)
+            const auth_id = supabaseReponse.data.user?.id;
+            if (auth_id == null) {
+                logger.info("User with supabase Id " + id + " not found")
                 return rep
                     .code(404)
                     .send()
-            }
-            else 
+            } else {
                 await AuthService.deleteUser(auth_id);
-            return await UserService.deleteUser(id);
+            }
+            logger.info("Successfully delete supabase user")
+            logger.info("Trying to delete user from database")
+            return await UserService.deleteUserById(id);
         } catch (err) {
             logger.info(err);
             return handleError(err, rep)
         }
     })
 
-    server.put(BASE_URL + '/:id', async (req: FastifyRequest<{Params: {id: number}}> , rep) => {
+    server.put(BASE_URL + '/:id', async (req: FastifyRequest<{ Params: { id: number } }>, rep) => {
         try {
             const id = req.params.id
             parseId(id)
@@ -74,6 +85,6 @@ export default async function userController(fastify: FastifyInstance, opts: any
             logger.error(err)
             return handleError(err, rep)
         }
-        
+
     })
 }
